Split the models file up even more
