+++ internal/fleet/output/models_kafka_test.go
